Single page app
This commit rewrites the front-end to comprise of a single HTML page
which updates in realtime as the user interacts with the page.

All of the applications endpoints now return JSON except `root.index`
which serves up the initial page markup, and `root.get_media` which
serves up a media file over the 206 partial media protocol.

There are three new JavaScript dependencies:

    angular                  # Synchronize data via the client
    angular-loading-bar      # No frills loading /progress bar
    angular-lazy-img         # No frills lazy image loading on demand